test: separate out nested struct tests
