+++ beet-solana/test/gpa.ts
@@ -1,47 +1,9 @@
 import test from 'tape'
 import spok from 'spok'
-import {
-  BeetStruct,
-  fixedSizeUtf8String,
-  i16,
-  i32,
-  u16,
-} from '@metaplex-foundation/beet'
+import { BeetStruct } from '@metaplex-foundation/beet'
 import { u8 } from '@metaplex-foundation/beet'
-import {
-  GetProgramAccountsConfig,
-  GetProgramAccountsFilter,
-  MemcmpFilter,
-  PublicKey,
-} from '@solana/web3.js'
-import { GpaBuilder } from '../src/gpa'
-import base58 from 'bs58'
-import { dbg } from './utils'
-
-const PROGRAM_ID = new PublicKey('cndy3Z4yapfJBmL3ShUp5exZKqR3z33thTzeNMm2gRZ')
-
-function isMemcmpFilter(x: GetProgramAccountsFilter): x is MemcmpFilter {
-  return (x as MemcmpFilter).memcmp != null
-}
-
-function withDecodedBytes(config: GetProgramAccountsConfig) {
-  const filters = []
-  if (config.filters != null) {
-    for (const x of config.filters) {
-      if (isMemcmpFilter(x)) {
-        filters.push({
-          memcmp: {
-            offset: x.memcmp.offset,
-            bytes: base58.decode(x.memcmp.bytes),
-          },
-        })
-      } else {
-        filters.push({ dataSize: x.dataSize })
-      }
-    }
-  }
-  return { ...config, filters }
-}
+import { PROGRAM_ID, Results, withDecodedBytes } from './utils'
+import { GpaBuilder } from 'src/beet-solana'
 
 test('gpa: fixed struct with one u8', (t) => {
   type Args = {
@@ -128,44 +90,6 @@ test('gpa: fixed struct with two u8s', (t) => {
   t.end()
 })
 
-type ResultsArgs = Pick<Results, 'win' | 'totalWin' | 'losses'>
-class Results {
-  constructor(
-    readonly win: number,
-    readonly totalWin: number,
-    readonly losses: number
-  ) {}
-
-  static readonly struct = new BeetStruct<Results, ResultsArgs>(
-    [
-      ['win', u8],
-      ['totalWin', u16],
-      ['losses', i32],
-    ],
-    (args) => new Results(args.win!, args.totalWin!, args.losses!),
-    'Results'
-  )
-}
-
-type TraderArgs = Pick<Trader, 'name' | 'results' | 'age'>
-class Trader {
-  constructor(
-    readonly name: string,
-    readonly results: Results,
-    readonly age: number
-  ) {}
-
-  static readonly struct = new BeetStruct<Trader, TraderArgs>(
-    [
-      ['name', fixedSizeUtf8String(4)], // offset: 0 (size: 2 * 4)
-      ['results', Results.struct], // offset: 8 (size: 7)
-      ['age', u8], // offset: 8 + 7 = 15
-    ],
-    (args) => new Trader(args.name!, args.results!, args.age!),
-    'Trader'
-  )
-}
-
 test('gpa: fixed struct with three ints', (t) => {
   let gpaBuilder: GpaBuilder<Results> = GpaBuilder.fromStruct<Results>(
     PROGRAM_ID,
@@ -201,78 +125,6 @@ test('gpa: fixed struct with three ints', (t) => {
       { memcmp: { offset: 1, bytes: Buffer.from([8]) } },
       { memcmp: { offset: 3, bytes: Buffer.from([-7]) } },
     ],
-  })
-
-  t.end()
-})
-
-test.only('gpa: fixed struct nested inside fixed struct', (t) => {
-  // Expected Filters
-  const nameFilter = (name: string) => ({
-    memcmp: {
-      offset: 0,
-      bytes: Buffer.concat([
-        Buffer.from([4, 0, 0, 0]), // length
-        Buffer.from(name),
-      ]),
-    },
-  })
-  const ageFilter = (age: number) => ({
-    memcmp: { offset: 15, bytes: Buffer.from([age]) },
-  })
-
-  const resultsFilter = (results: Results) => {
-    const bytes = Results.struct.serialize(results)[0]
-    return {
-      memcmp: {
-        offset: 8,
-        bytes,
-      },
-    }
-  }
-
-  // Prep
-  let gpaBuilder: GpaBuilder<Trader> = GpaBuilder.fromStruct<Trader>(
-    PROGRAM_ID,
-    Trader.struct
-  )
-  function prepCase(comment: string) {
-    t.comment(comment)
-    gpaBuilder = GpaBuilder.fromStruct<Trader>(PROGRAM_ID, Trader.struct)
-  }
-
-  // Tests
-  prepCase(
-    `name whic is before results struct - gpaBuilder.addFilter('name', 'trad')`
-  )
-  gpaBuilder.addFilter('name', 'trad')
-  spok(t, withDecodedBytes(gpaBuilder.config), {
-    filters: [nameFilter('trad')],
-  })
-
-  prepCase(
-    `age which is after results struct - gpaBuilder.addFilter('age', 99)`
-  )
-  gpaBuilder.addFilter('age', 99)
-  spok(t, withDecodedBytes(gpaBuilder.config), {
-    filters: [ageFilter(99)],
-  })
-
-  prepCase(`gpaBuilder.addFilter('name', 'trad').addFilter('age', 99)`)
-  gpaBuilder.addFilter('name', 'trad').addFilter('age', 99)
-  spok(t, withDecodedBytes(gpaBuilder.config), {
-    filters: [nameFilter('trad'), ageFilter(99)],
-  })
-
-  prepCase(`gpaBuilder.addFilter('results', { win, totalWin, losses })`)
-  const results = {
-    win: 3,
-    totalWin: 4,
-    losses: -100,
-  }
-  gpaBuilder.addFilter('results', results)
-  spok(t, withDecodedBytes(gpaBuilder.config), {
-    filters: [resultsFilter(results)],
   })
 
   t.end()
