+++ beet-solana/test/utils/index.ts
@@ -1,4 +1,22 @@
+import {
+  BeetStruct,
+  fixedSizeUtf8String,
+  i32,
+  u16,
+  u8,
+} from '@metaplex-foundation/beet'
+import {
+  GetProgramAccountsConfig,
+  GetProgramAccountsFilter,
+  MemcmpFilter,
+  PublicKey,
+} from '@solana/web3.js'
+import base58 from 'bs58'
 import * as util from 'util'
+
+export const PROGRAM_ID = new PublicKey(
+  'cndy3Z4yapfJBmL3ShUp5exZKqR3z33thTzeNMm2gRZ'
+)
 
 export function inspect(obj: any, colors = true) {
   console.log(util.inspect(obj, { colors, depth: 5 }))
@@ -7,4 +25,68 @@ export function inspect(obj: any, colors = true) {
 export function dbg(obj: any) {
   inspect(obj)
   return obj
+}
+
+export function isMemcmpFilter(x: GetProgramAccountsFilter): x is MemcmpFilter {
+  return (x as MemcmpFilter).memcmp != null
+}
+
+export function withDecodedBytes(config: GetProgramAccountsConfig) {
+  const filters = []
+  if (config.filters != null) {
+    for (const x of config.filters) {
+      if (isMemcmpFilter(x)) {
+        filters.push({
+          memcmp: {
+            offset: x.memcmp.offset,
+            bytes: base58.decode(x.memcmp.bytes),
+          },
+        })
+      } else {
+        filters.push({ dataSize: x.dataSize })
+      }
+    }
+  }
+  return { ...config, filters }
+}
+
+// -----------------
+// Sample Structs
+// -----------------
+export type ResultsArgs = Pick<Results, 'win' | 'totalWin' | 'losses'>
+export class Results {
+  constructor(
+    readonly win: number,
+    readonly totalWin: number,
+    readonly losses: number
+  ) {}
+
+  static readonly struct = new BeetStruct<Results, ResultsArgs>(
+    [
+      ['win', u8],
+      ['totalWin', u16],
+      ['losses', i32],
+    ],
+    (args: ResultsArgs) => new Results(args.win!, args.totalWin!, args.losses!),
+    'Results'
+  )
+}
+
+export type TraderArgs = Pick<Trader, 'name' | 'results' | 'age'>
+export class Trader {
+  constructor(
+    readonly name: string,
+    readonly results: Results,
+    readonly age: number
+  ) {}
+
+  static readonly struct = new BeetStruct<Trader, TraderArgs>(
+    [
+      ['name', fixedSizeUtf8String(4)], // offset: 0 (size: 2 * 4)
+      ['results', Results.struct], // offset: 8 (size: 7)
+      ['age', u8], // offset: 8 + 7 = 15
+    ],
+    (args) => new Trader(args.name!, args.results!, args.age!),
+    'Trader'
+  )
 }
